Build an extra profile which can load in a non-browser context.
The only requirement for this context is that it provides an object called
'globals' where the 'angular' object will be installed, instead of
installing it on 'window' as we'd do in the normal browser-oriented
version.

The generated module still won't be completely functional because certain
services in the 'ng' module expect to have a browser available. The goal
here is just to make the module loadable so that the caller can then
provide its own overridden versions of the browser-oriented services
that make more sense in the non-browser environment.